Lesson 4.1 Performance, TypeScript, Testing, Project Structure
